+++ backend/users/Touir.ExpensesManager.Users/Validators/RegisterRequestValidator.cs
@@ -7,9 +7,12 @@ namespace Touir.ExpensesManager.Users.Validators
     {
         public RegisterRequestValidator()
         {
-            RuleFor(x => x.FirstName).NotEmpty().WithMessage("MISSING_PARAMETERS");
-            RuleFor(x => x.LastName).NotEmpty().WithMessage("MISSING_PARAMETERS");
-            RuleFor(x => x.Email).NotEmpty().WithMessage("MISSING_PARAMETERS");
+            RuleFor(x => x.FirstName).NotEmpty().WithMessage("MISSING_PARAMETERS")
+                .MaximumLength(100).WithMessage("FIELD_TOO_LONG");
+            RuleFor(x => x.LastName).NotEmpty().WithMessage("MISSING_PARAMETERS")
+                .MaximumLength(100).WithMessage("FIELD_TOO_LONG");
+            RuleFor(x => x.Email).NotEmpty().WithMessage("MISSING_PARAMETERS")
+                .MaximumLength(100).WithMessage("FIELD_TOO_LONG");
         }
     }
 }